Implement a bytecode reader.
We no longer need to hardcode the program.
The start function reads a file into Read/Write memory performing at
least one allocation, then it replaces all opcodes with the operation
pointers.

This startup time is about 0.1ms and shouldn't be problematic unless
for very large programs. The input file size is currently capped
at 1GB, a mostly arbitrary number.

Returning errors from start to main proved to be too much work, so I
opted instead for printing a message to stderr and exit with code 1,
the good 'ol C way.

I encountered some alignment issues with the input variable, so I had
to assert to Zig that there was no issue using @alignCast, the
correctness of which remains to be verified.