Add quick and dirty assembler in Python.
Pylint thinks my Location class has "too few public methods". What
a time to be writing code.

match/case syntax was very pleasant to use and is more powerful than I
previously thought.

This assembler doesn't check if operands should be immediate or
registers, as that's too much logic for now. Maybe a declarative
approach would be much more fruitful here.

There is also the issue of endianness, which I hardcoded to Little at
least for now.

OPCODE_NUMBER is a map from opcode names in code to integer values,
and will sadly have to be updated manually as the bytecode changes.